feat: ISSUE_TEMPLATEを追加 (#3)
* feat: ISSUE_TEMPLATEを追加

* fix: ghlmに名前変更してリファクタリング

* fix: config.yamlのdefaultsフィールドを削除

* feat: pr templateを追加

* feat: テストワークフローとAIメンター化設定

* Revert "feat: テストワークフローとAIメンター化設定"

This reverts commit 019c2850d3fa12059c86bcdc37554a001e67f274.

* fix: 不要なキャストを削除

+++ src/models.rs
@@ -25,7 +25,7 @@ impl From<octocrab::models::Repository> for Repository {
                 .owner
                 .map_or_else(|| "N/A".to_string(), |owner| owner.login),
             description: repo.description,
-            stars: repo.stargazers_count.unwrap_or(0) as u32,
+            stars: repo.stargazers_count.unwrap_or(0),
             private: repo.private.unwrap_or(false),
         }
     }
